Move config component to under the setting dir
